remove group_by from function registry
